feat: Add spawnChildJobsToWaitOn function for creating child jobs
This commit adds a new function `spawnChildJobsToWaitOn` that allows creating child jobs to wait on. The function can handle both single and multiple child jobs and adds the parent job information to the child jobs. This function is used in the `createWorkerMainFunction` to spawn and add child jobs to the FlowProducer.