aider: The migration has been modified to add a uniqueness constraint to the `jobs_log.job_id` column by adding a unique index to the `job_id` column in the `jobs_log` table.
# Aider chat conversation:

ASSISTANT: To modify the `20231111215102_create_job_deps_table.js` migration so that it adds a uniqueness constraint to `jobs_log.job_id`, we will need to add a unique index to the `job_id` column in the `jobs_log` table. This change should be made in the `up` function to add the constraint and in the `down` function to remove it when rolling back the migration.

Here are the changes for the `20231111215102_create_job_deps_table.js` file:

```javascript
migrations/20231111215102_create_job_deps_table.js
<<<<<<< SEARCH
exports.up = function (knex) {
  // create dependency table with parent and child job ids referencing the jobs_log table
  return knex.schema.createTable("job_deps", (table) => {
    table.primary(["project_id", "parent_job_id", "child_job_id"]);
    table.string("project_id");
    table.string("parent_job_id").notNullable();
    table.string("child_job_id").notNullable();
    table.foreign("parent_job_id").references("jobs_log.job_id");
    table.foreign("child_job_id").references("jobs_log.job_id");
    table.unique(["project_id", "parent_job_id", "child_job_id"]);
  });
};
=======
exports.up = function (knex) {
  return knex.schema
    .createTable("job_deps", (table) => {
      table.primary(["project_id", "parent_job_id", "child_job_id"]);
      table.string("project_id");
      table.string("parent_job_id").notNullable();
      table.string("child_job_id").notNullable();
      table.foreign("parent_job_id").references("jobs_log.job_id");
      table.foreign("child_job_id").references("jobs_log.job_id");
      table.unique(["project_id", "parent_job_id", "child_job_id"]);
    })
    .then(() => knex.schema.alterTable("jobs_log", (table) => {
      table.unique("job_id");
    }));
};
>>>>>>> REPLACE

<<<<<<< SEARCH
exports.down = function (knex) {
  return knex.schema.dropTable("job_deps");
};
=======
exports.down = function (knex) {
  return knex.schema
    .dropTable("job_deps")
    .then(() => knex.schema.alterTable("jobs_log", (table) => {
      table.dropUnique("job_id");
    }));
};
>>>>>>> REPLACE
```

These changes will ensure that when the `up` migration is run, a unique constraint is added to the `job_id` column of the `jobs_log` table. When the `down` migration is run, this unique constraint is removed.