aider: Add Neon crate as a dependency in Cargo.toml.
# Aider chat conversation:

USER: I added these *read-write* files: packages/shared-crosslang/Cargo.toml

ASSISTANT: To enable the Rust code in `def_graph.rs` to be compiled into a native Node.js addon using Neon, we need to add Neon as a dependency in the `Cargo.toml` file. Here's the diff for the changes:

```diff
--- packages/shared-crosslang/Cargo.toml
+++ packages/shared-crosslang/Cargo.toml
@@ ... @@
 [dependencies]
 # Default enable napi4 feature, see https://nodejs.org/api/n-api.html#node-api-version-matrix
 napi = { version = "2.12.2", default-features = false, features = ["napi4"] }
 napi-derive = "2.12.2"
 petgraph = "0.6.4"
+neon = "0.10.0"

 [build-dependencies]
 napi-build = "2.0.1"
```

This change adds the `neon` crate to the list of dependencies, which will allow you to use Neon's features to create the Node.js addon. Make sure to run `cargo build` after updating the `Cargo.toml` file to download and compile the new dependency.