aider: Added unit test for the Rust version of `ensure_node` method in `def_graph.rs` with necessary test structure and `assert_matches` dependency.
# Aider chat conversation:

USER: add a unit test for the RS version of ensure_node that is similar to the test for the TS version of ensureNode. Create necessary unit test structure because there isn't one yet.

ASSISTANT: To add a unit test for the Rust version of `ensure_node`, we need to create a new test module within the `def_graph.rs` file. Since there isn't an existing test structure, we will also need to add the necessary configuration to the `Cargo.toml` file to include the test dependencies.

Here are the diffs for these changes:

```diff
--- crosslang/shared/Cargo.toml
+++ crosslang/shared/Cargo.toml
@@ ... @@
 [dependencies]
 petgraph = "0.6.0"
+
+[dev-dependencies]
+assert_matches = "1.5.0"
```

```diff
--- crosslang/shared/src/systems/def_graph.rs
+++ crosslang/shared/src/systems/def_graph.rs
@@ ... @@
 use petgraph::graph::{DiGraph, NodeIndex};
 use std::collections::HashMap;
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use assert_matches::assert_matches;
+
+    #[test]
+    fn test_ensure_node() {
+        let mut graph = DefGraph::<&str>::new();
+
+        // Ensure a node is created
+        let node_id = graph.ensure_node("TestNode", "TestData");
+        assert_matches!(graph.graph.node_weight(node_id), Some(&"TestData"));
+
+        // Ensure the same node is retrieved with the same ID
+        let same_node_id = graph.ensure_node("TestNode", "TestData");
+        assert_eq!(node_id, same_node_id);
+
+        // Ensure the node data is not overwritten if the same ID is used
+        let same_node_id_with_different_data = graph.ensure_node("TestNode", "DifferentData");
+        assert_matches!(graph.graph.node_weight(same_node_id_with_different_data), Some(&"TestData"));
+    }
+}
```

These changes add the `assert_matches` crate as a development dependency for writing more expressive tests. In the `def_graph.rs` file, a new test module is created with a single test function `test_ensure_node` that checks the behavior of the `ensure_node` method. The test ensures that a node is created with the given ID and data, and if the same ID is used again, the existing node is retrieved without overwriting its data.